Improve the arrays of Cartesian coordinates
The arrays containing the transformed Cartesian coordinates now have
useful dimensions and are coordinates of the datasets.  This will
enable easier plotting of the data.

Furthermore, a new array with the name "mask" is created, so that it
is not necessary anymore to search for NaNs in the (salinity) data.