Add smoothing code and preliminary slope analysis
The new library tools.smoothing is added, which provides a smoothing algorithm for 2D arrays.

This algorithm is used in the notebook for the topography creation to make a smoother bathymetry, but this smoothed topography is currently not employed.  In a new part of the same notebook, the local slopes of the original and smoothed topographies are calculated and displayed.