Activate ice tongue, subglacial runoff and topography smoothing
A few lines of code are added to smooth the ice thickness and to use the smoothed data in the GETM topography. As an attempt to improve model stability, the shape of an island in the buffer zone (halo) of the open boundary is simplified. A bug in the indices of subglacial runoff is corrected (GETM indices start with 1, not 0).

The ice tongue and the subglacial runoff are now activated in the setup.