Add code to create 3D boundary conditions
The notebook for boundary conditions now creates the 3D conditions
(temperature, salinity) in addition to the 2D conditions (elevation).
The former content of the script that creates the stratification files
is absorbed into this notebook, since most of the code in the script is
needed in the notebook, and also since the initial stratification should
correspond to the boundary conditions.

In the configuration, the 3D boundaries are activated and the duration
of the simulation is now extended to 1 year, because this information
is used to create the boundary conditions.  In addition, some more
variables are set to their intended (final?) values.