Determine the grounding line just before saving
Since the grounding line could still change by masking grid points with
grounded ice, it is important to determine the indices of the grounding
line points only when the topography is finshed, right before saving it.

This commit also introduces a new check to make sure that the size
of the topography fits the values given in the GETM configuration.