Implement the 500 m resolution topography
In this commit, the topography notebook and the configuration are
adjusted to create the 500 m resolution (instead of the 1 km resolution
that was used before).  For simplicity, the code for the lower
resolution was removed, i.e., replaced with the new code.  The old
implementation still exists in its own branch and in release v0.1.0.