Adjust boundaries and split factor for 500 m resolution
Compared to the 1 km-setup, the southern boundary is simpler for the higher resolution 500 m-topography, because the boundary is not interrupted by an island. As there are more grid points, only every 20th (instead of every 10th) longitude is labeled to prevent overlapping labels.

Since the 2D time step was previously reduced and stability violations in the baroclinic mode are not expected, the 3D time step is increased.