Update the flexible output configuration
A new output file contains all time-independent geometric information,
so there's no need to save metric and domain information in every file.
While this data is not large, it clutters the namespace.

Snapshots are divided into hourly output for 2D quantities and daily
output for 3D quantities.  This gives the overall development in a nice
temporal resolution without requiring too much storage.

Time-averaged quantities are only saved at the end, because they are not
needed in the spin-up and only slow-down the model run.

Velocity fields are no longer renamed, since this can cause confusion.

The explanations at the end of the file are improved.