Use numerical diffusion An proportional to sqrt(H)
To get rid of small-amplitude noise in the sea surface elevation, we
enable horizontal numerical diffusion.  We choose the diffusion
coefficient A_n to be proportional to sqrt(H), where H is the bottom
depth.  The max value of A_n is 50, as used by Reinert et al. (2023),
which turns out to be sufficient also in this setup.

Also, we extend the duration of the simulation to 2 years, so that the
model reaches a quasi-steady state.

For consistency with GETM output, long names of variables are not
capitalized anymore in the topography dataset.