Implement realistic time-mean boundary conditions
Temperature and salinity data from a global ocean model are used as
initial and boundary conditions for the setup.  The data is averaged in
time, so that the boundary conditions are not time-dependent, and a
quasi-steady state can be reached.  The initial and boundary conditions
are equal on boundary points, so that the initial situation is stable.

The previously incomplete notebook to interpolate global model data onto
the GETM grid for boundary conditions is now re-purposed into a notebook
to create initial conditions and has been completed. The notebook making
boundary conditions uses these initial conditions,  and does not require
anymore the specification of a stratification in the configuration file.