Improve the mask near the grounding line
This commit introduces some fine-tuning of the mask, in particular near
the grounding line and along the fjord walls.  These areas are tricky,
since the two topography datasets do not match perfectly and are to
some extent interpolated there.  With careful handling of the data in
these places and attention to details, a better land/ocean mask and
a better topography are obtained.

In the notebook for initial conditions, a comment is adjusted to the new
topography and two minor improvements are introduced.